requirements.txt
# Core
python-dotenv>=1.0.0

# Crawling & Ingestion
crawl4ai>=0.4.0
langchain-text-splitters>=0.3.0

# Vector DB
chromadb>=0.4.24

# API
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# LLM
openai>=1.6.0

# UI
streamlit>=1.31.0
requests>=2.31.0
